Fix SocketEx

+++ Network/SocketEx/app/src/main/java/edu/yujie/socketex/util/FileExt.kt
@@ -2,28 +2,9 @@ package edu.yujie.socketex.util
 
 import java.io.File
 
-fun File.createFile(): File {
-    if (exists()) delete()
-    createNewFile()
-    return this
+fun File.createFile(fileName: String): File {
+    val file = File(this, fileName)
+    if (file.exists()) file.delete()
+    file.createNewFile()
+    return file
 }
-
-object FileExt {
-    fun createFile(filePath: File?, fileName: String): File {
-        val file = File(filePath, fileName)
-        if (file.exists()) {
-            file.delete()
-        }
-        file.createNewFile()
-        return file
-    }
-}
-
-//fun File.createFile(): BehaviorSubject<File> {
-//    val subject = BehaviorSubject.create<File>()
-//    if (exists()) delete()
-//    createNewFile()
-//    subject.onNext(this)
-//    return subject
-//}
-
